alx_travel_app_0x00/alx_travel_app/requirement.txt
# Core Django and REST Framework
Django>=4.0,<5.0
djangorestframework>=3.14

# CORS handling
django-cors-headers>=4.0.0

# MySQL database adapter
mysqlclient>=2.2.0

# Swagger / OpenAPI documentation
drf-yasg>=1.21.5

# Asynchronous task queue
celery>=5.3.1

# RabbitMQ is external, but you need a client to communicate (optional)
kombu>=5.3.0